put fault data structures in errorGuage.py
